feat: hide handlers when focus outside + make a standalone component for the handlers
