+++ src/components/toolbar/index.js
@@ -41,6 +41,7 @@ import {
 	__experimentalColumnMover,
 	__experimentalRowSettings,
 	__experimentalColumnSize,
+	__experimentalColumnsHandlers,
 } from './components';
 
 /**
@@ -306,7 +307,11 @@ const MaxiToolbar = props => {
 								setAttributes({ rowPattern })
 							}
 							breakpoint={deviceType}
+						/>
+
+						<__experimentalColumnsHandlers
 							toggleHandlers={toggleHandlers}
+							blockName={name}
 						/>
 						<Link
 							blockName={name}
